content_creation_agency/youtube_analyzer/tools/channel_analyzer.py
from agency_swarm.tools import BaseTool
from pydantic import Field
import os
from dotenv import load_dotenv
import json
from googleapiclient.discovery import build
from googleapiclient.errors import HttpError

load_dotenv()

class channel_analyzer(BaseTool):
    """
    Analyzes YouTube channel performance and metrics.
    """
    channel_id: str = Field(
        ..., 
        description="The YouTube channel ID to analyze"
    )
    metrics: list = Field(
        default=["subscribers", "views", "engagement"],
        description="List of metrics to analyze"
    )

    def run(self):
        """
        Analyzes the specified YouTube channel's performance metrics.
        """
        # TODO: Implement actual channel analysis logic
        # This is a placeholder implementation
        return f"Analysis of channel {self.channel_id} for metrics: {', '.join(self.metrics)}"

if __name__ == "__main__":
    tool = channel_analyzer(
        channel_id="UC1234567890",
        metrics=["subscribers", "views"]
    )
    print(tool.run()) 