content_creation_agency/requirements.txt
agency-swarm>=0.1.0
python-dotenv>=0.19.0
openai>=1.0.0
nltk>=3.8.1
pandas>=2.0.0
google-api-python-client>=2.0.0
tavily-python>=0.1.0
pytrends>=4.9.0
vaderSentiment>=3.3.2
gradio>=4.0.0 